+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/LanguageServices/InitializeParameter/IInitializeParameterService.cs
@@ -13,6 +13,8 @@ namespace Microsoft.CodeAnalysis.InitializeParameter;
 
 internal interface IInitializeParameterService : ILanguageService
 {
+    bool IsThrowNotImplementedProperty(Compilation compilation, IPropertySymbol property, CancellationToken cancellationToken);
+
     void InsertStatement(
         SyntaxEditor editor, SyntaxNode functionDeclaration, bool returnsVoid, SyntaxNode? statementToAddAfter, SyntaxNode statement);
